Fix code
Changed the method with the completion block.

DI, assembly and configuration of dependencies are done in one place. A separate class has been created.

APIArticManager.shared.getInfo - APIArticManager is closed by the protocol and injected through the initializer.

Removed Base VC.

The entire layout is rendered in ArticView and installed in loadView()